divide NAV REL POD NED into modules per protocol version and share common flags
